requirements.txt
absl-py==1.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
Cython==0.29.30
# d4rl @ git+https://github.com/rail-berkeley/d4rl@d842aa194b416e564e54b0730d9f934e3e32f854
# dm-control @ git+https://github.com/deepmind/dm_control@41d0c7383153f9ca6c12f8e865ef5e73a98759bd
dm-env==1.5
dm-tree==0.1.7
fasteners==0.17.3
fonttools==4.33.3
glfw==2.5.3
# gym==0.24.1
gym-notices==0.0.7
h5py==3.7.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
kiwisolver==1.4.2
labmaze==1.0.5
lxml==4.9.0
matplotlib==3.5.2
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
# mujoco==2.2.0
# mujoco-py==2.0.2.13
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==4.21.1
pybullet==3.2.5
pycparser==2.21
PyOpenGL==3.1.6
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
scipy==1.8.1
six==1.16.0
termcolor   #==1.1.0
# torch==1.10.1+cu111
# torchaudio==0.10.1+rocm4.1
# torchvision==0.11.2+cu111
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
